problem/api/requirements.txt
flask==3.0.3
sqlalchemy==2.0.36
python-dateutil==2.9.0.post0
pytz==2024.1
pytest==8.3.3
